Default cookie expiration should be in the future
In the current implementation, a cookie on a response gets a default
expiration time when put into a CookieJar defined by this code snippet:

    getExpiryTime Nothing Nothing= UTCTime (ModifiedJulianDay 0) (secondsToDiffTime 0)

This results in a default expiration time of 1858-11-17T00:00Z - which
is in the past.  As a result, this cookie will be excluded from any
requests that are made with the same CookieJar attached.

The default expiration time should be in the future, so that cookies
that do not specify an expiration time or a max age are included in
future requests.  RFC 6265 specifies that when no expiration time or max
age is given, "Set the cookie's expiry-time to the latest representable
date."  The UTCTime type uses an Integer to represent the date, meaning
that there is no latest representable date.  So we have to pick an
arbitrary date in the future for default expiration times.

This commit changes the default expiration time logic to assign a date
that is approximately 1000 years in the future:

    getExpiryTime Nothing Nothing = UTCTime (365000 `addDays` utctDay now) (secondsToDiffTime 0)

Running this on 2012-02-21 results in the time 3011-6-24T00:00Z.