Separate tests not requiring internet access.
Most of the tests in the existing test suite do not require internet
access. However a few of them do make requests to external hosts --
this causes the test suite to fail when the user's network connection
is down or otherwise unavailable.

This commit separates the existing "spec" test suite into "spec" and
"spec-nonet" components, the latter being the subset of tests that
work when the network connection is down. A separate "test-nonet"
directory is added to preserve the existing hspec autodiscovery magic.

Some redundant imports were removed after the restructuring, and a few
occurrences of unused variables were changed to "_".