Allow HTTPS proxies
This commit fixes a bug where specifying a proxy prefixed by 'https'
caused an HttpException to be raised. This seems like a scheme that
should be valid, particularly since we're allowing an `https_proxy` to
be specified in the environment.

With this change I also added test cases for previous behavior that
wasn't tested, i.e., 'http:' should be a valid proxy uri scheme name,
and 'ftp:' should not be a valid proxy uri scheme name.