Don't modify request n+1 times
There was a use case where a colleague created a `Manager` to send
authorized requests by appending the header field `Authorization: Bearer
...` to the list of a request header fields list, using the
`managerModifyRequest` setting. The server replied with 401 errors even
if the token was valid. The reason is that `managerModifyRequest`
function was applied twice and the server was interpreting the duplicate
`Authorization` header field as a value that is defined as a
comma-separated list. According to section 4.2 Message Headers of RFC
2616:

> Multiple message-header fields with the same field-name MAY be
> present in a message if and only if the entire field-value for that
> header field is defined as a comma-separated list [i.e., #(values)].
> It MUST be possible to combine the multiple header fields into one
> "field-name: field-value" pair, without changing the semantics of the
> message, by appending each subsequent field-value to the first, each
> separated by a comma. The order in which header fields with the same
> field-name are received is therefore significant to the
> interpretation of the combined field value, and thus a proxy MUST NOT
> change the order of these field values when a message is forwarded.

In `responseOpen` the first application of `getModifiedRequestManager`
is useful to pass a consistent request value to `wrapExc` and to pass
the correct value of redirect count to the first application of
`httpRedirect'` in `go manager (redirectCount req0) req0`, but I think
that using `req0` as the last argument was redundant because `go`
applies again `getModifiedRequestManager` in the block that is used as
the second parameter of `httpRedirect'`.

Note about the new test case: I added a test that sends a request to
httpbin.org/headers and parses the response to check that the client
doesn't send the same header n+1 times. Unfortunately httpbin.org
mistakenly removes duplicate header fields, see
https://github.com/kennethreitz/httpbin/issues/355, so I had to use
request's `redirectCount` as an accumulator to send two different header
fields.