avoid throwing exception on a malformed HTTP header
A line in the HTTP response that lacks a ':' caused an exception,
but other commonly used http implementations (curl, wget, firefox)
are robust in what they accept here. So a web server that is
misconfigured with eg, a '\n' in the middle of a header value,
would seem to work ok generally but break when http-client is used.
Fixed by ignoring such bogus and unparseable lines.

The InvalidHeader constructor is no longer used in http-client, so could
be later removed; it has been left in to avoid this bug fix getting
entangled with an API bump.

Fixes https://github.com/snoyberg/http-client/issues/398