remove original request from Response
Deliver it instead by way of the Manager's response modifier.

See https://github.com/snoyberg/http-client/pull/464

This patch accomplishes the same thing as that one in essentially the
same way, but improves upon it by not requiring a request to exist in
order to construct a response. This can be useful in e.g. testing /
mocking definitions in which we want to express a base fixed response.
In other wrods, it is "less breaky" and easier to adopt.