http-client: Avoid a space leak when storing original requests
`Network.HTTP.Client.Response.getResponse` stores the original request
inside a `Response`, but attempts to override its request body with an
empty one. This record update was lazy, and a reference to the
original request body would persist, causing a space leak.

Force the response to WHNF, and use bang patterns to force the Request
and ultimately its body to WHNF too.